+++ src/main/java/pl/futurecollars/invoicing/service/InvoiceService.java
@@ -17,11 +17,11 @@ public class InvoiceService {
         this.db = database;
     }
 
-    public int save(Invoice invoice) {
+    public long save(Invoice invoice) {
         return db.save(invoice);
     }
 
-    public Optional<Invoice> getById(int id) {
+    public Optional<Invoice> getById(long id) {
         return db.getById(id);
     }
 
@@ -29,7 +29,7 @@ public class InvoiceService {
         return db.getAll();
     }
 
-    public Optional<Invoice> update(int id, Invoice updatedInvoice) {
+    public Optional<Invoice> update(long id, Invoice updatedInvoice) {
         if (inMemoryDatabase.getInvoiceInMemoryDatabase().containsKey(id)) {
             return Optional.empty();
         } else {
@@ -37,7 +37,7 @@ public class InvoiceService {
         }
     }
 
-    public Optional<Invoice> delete(int id) {
+    public Optional<Invoice> delete(long id) {
         return db.delete(id);
     }
 }
